pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "phishkiller"
version = "0.1.0"
description = "Phishing kit tracking and analysis platform"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "certstream>=1.12",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "typer[all]>=0.15.0",
    "python-dotenv>=1.0.0",
    "stix2>=3.0.1",
    "python-multipart>=0.0.18",
    "rarfile>=4.2",
    "python-Levenshtein>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "httpx",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
tlsh = [
    "py-tlsh>=4.7.2",
]
yara = [
    "yara-python>=4.5.1",
]

[project.scripts]
phishkiller = "phishkiller.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
